app: Restructure things
