+++ config/README
