+++ config/gulp/gulpfile.js
@@ -4,7 +4,6 @@ import { register, createPackager } from './utils'
 
 register(gulp);
 
-// gulp.task('default', 'build');
 
 /* Build Tasks */
 gulp.task('build', gulp.series(gulp.parallel('build:app', 'build:electron')));
@@ -21,11 +20,10 @@ gulp.task('package', gulp.series('pre-package', createPackager("current")));
 gulp.task('package:win', gulp.series('pre-package', createPackager("windows")));
 gulp.task('package:mac', gulp.series('pre-package', createPackager("mac")));
 gulp.task('package:linux', gulp.series('pre-package', createPackager("linux")));
-gulp.task('package:appveyor', gulp.series('pre-package', createPackager("appveyor")));
-gulp.task('package:travis', gulp.series('pre-package', createPackager("travis")));
 gulp.task('package:portable', gulp.series('pre-package', createPackager("current portable")));
-// architectures: ia32, x64, armv7l
-// platforms:  linux, win32, darwin, mas
+
+gulp.task('publish:appveyor', gulp.series('pre-package', createPackager("appveyor")));
+gulp.task('publish:travis', gulp.series('pre-package', createPackager("travis")));
 
 
 /* Helper Tasks */
