+++ config/webpack/ConsoleClearPlugin.js
