+++ config/webpack/aliases.js
