+++ config/webpack/loaders.js
