+++ config/webpack/vendors.js
