+++ config/webpack/webpack.dev.js
@@ -3,13 +3,13 @@ import path from 'path';
 
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import WebpackBuildNotifierPlugin from 'webpack-build-notifier';
-import ConsoleClearPlugin from '../ConsoleClearPlugin';
+import ConsoleClearPlugin from './ConsoleClearPlugin';
 
 import config  from 'config';
 import paths from 'config/paths';
-import createLoaders from '../loaders';
-import aliases from '../aliases';
-import vendors from '../vendors';
+import createLoaders from './loaders';
+import aliases from './aliases';
+import vendors from './vendors';
 
 
 module.exports = {
@@ -36,7 +36,7 @@ module.exports = {
         alias: aliases,
         modules: ['node_modules', config.paths.app_modules, config.paths.app]
     },
-    module: { loaders: createLoaders("production") },
+    module: { loaders: createLoaders("development") },
     devServer: {
         contentBase: config.paths.build,
         noInfo: false,
@@ -45,17 +45,12 @@ module.exports = {
         inline: true,
         port: config.server.port,
         host: config.server.host,
-        stats: {maxModules: 0}, // disable modules
-        headers: {
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
-            "Access-Control-Allow-Headers": "X-Requested-With, content-type, Authorization"
-        }
+        stats: { maxModules: 0 }, // disable modules
     },
     plugins: [
         new WebpackBuildNotifierPlugin(),
-        new webpack.NoEmitOnErrorsPlugin(),
         new webpack.NamedModulesPlugin(),
+        new webpack.NoEmitOnErrorsPlugin(),
         new webpack.HotModuleReplacementPlugin(),
         // new webpack.optimize.CommonsChunkPlugin({
         //     name: "vendor",
@@ -63,7 +58,7 @@ module.exports = {
         //     children: true,
         // }),
         new webpack.EnvironmentPlugin({
-            NODE_ENV: 'production', // uses 'development' unless process.env.NODE_ENV is defined
+            NODE_ENV: 'development', // uses 'development' unless process.env.NODE_ENV is defined
             DEBUG: true
         }),
         new webpack.ProvidePlugin({
@@ -72,8 +67,8 @@ module.exports = {
             "window.jQuery": "jquery"
         }),
         new HtmlWebpackPlugin({
-            template: config.paths.app_html,
-            inject: false
+            template: paths.app_html,
+            inject: 'body'
         }),
         new ConsoleClearPlugin() // custom plugin to clear the console before each bundle
     ],
