+++ config/webpack/webpack.preview.js
@@ -3,13 +3,13 @@ import path from 'path';
 
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import WebpackBuildNotifierPlugin from 'webpack-build-notifier';
-import ConsoleClearPlugin from '../ConsoleClearPlugin';
+import ConsoleClearPlugin from './ConsoleClearPlugin';
 
 import config  from 'config';
 import paths from 'config/paths';
-import createLoaders from '../loaders';
-import aliases from '../aliases';
-import vendors from '../vendors';
+import createLoaders from './loaders';
+import aliases from './aliases';
+import vendors from './vendors';
 
 
 module.exports = {
@@ -36,7 +36,7 @@ module.exports = {
         alias: aliases,
         modules: ['node_modules', config.paths.app_modules, config.paths.app]
     },
-    module: { loaders: createLoaders("development") },
+    module: { loaders: createLoaders("production") },
     devServer: {
         contentBase: config.paths.build,
         noInfo: false,
@@ -45,12 +45,17 @@ module.exports = {
         inline: true,
         port: config.server.port,
         host: config.server.host,
-        stats: { maxModules: 0 }, // disable modules
+        stats: {maxModules: 0}, // disable modules
+        headers: {
+            "Access-Control-Allow-Origin": "*",
+            "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
+            "Access-Control-Allow-Headers": "X-Requested-With, content-type, Authorization"
+        }
     },
     plugins: [
         new WebpackBuildNotifierPlugin(),
-        new webpack.NamedModulesPlugin(),
         new webpack.NoEmitOnErrorsPlugin(),
+        new webpack.NamedModulesPlugin(),
         new webpack.HotModuleReplacementPlugin(),
         // new webpack.optimize.CommonsChunkPlugin({
         //     name: "vendor",
@@ -58,7 +63,7 @@ module.exports = {
         //     children: true,
         // }),
         new webpack.EnvironmentPlugin({
-            NODE_ENV: 'development', // uses 'development' unless process.env.NODE_ENV is defined
+            NODE_ENV: 'production', // uses 'development' unless process.env.NODE_ENV is defined
             DEBUG: true
         }),
         new webpack.ProvidePlugin({
@@ -67,8 +72,8 @@ module.exports = {
             "window.jQuery": "jquery"
         }),
         new HtmlWebpackPlugin({
-            template: paths.app_html,
-            inject: 'body'
+            template: config.paths.app_html,
+            inject: false
         }),
         new ConsoleClearPlugin() // custom plugin to clear the console before each bundle
     ],
