+++ electron/events/handleBeforeSendHeaders.js
