+++ electron/events/handleRedirect.js
