+++ electron/events/handleSquirrel.js
