+++ electron/setup.js
