+++ electron/views/main.html
