+++ electron/views/preloader.html
@@ -97,7 +97,7 @@
     </style>
 </head>
 <body>
-    <img src="../../../public/images/logo.png">
+    <img src="../../public/images/logo.png">
     <div class="preload-text">Building<span>.</span><span>.</span><span>.</span></div>
     <div class="versions">
         <span class="title">
