+++ package.json
@@ -7,10 +7,10 @@
   "scripts": {
     "___CORE___": "These scripts will get you going with Lurka",
     "start": "npm run dev",
-    "dev": "concurrently -kr \"npm run frontend\" \"npm run backend\"",
-    "preview": "concurrently -kr \"npm run frontend:preview\" \"npm run backend\"",
+    "dev": "concurrently -kr \"npm run app\" \"npm run electron\"",
     "test": "npm run mocha",
     "test:watch": "npm run mocha -R min -w",
+    "preview": "concurrently -kr \"npm run app:preview\" \"npm run electron\"",
     "___BUILD___": "These scripts will bundle, distribute and publish Lurka",
     "build": "npm run gulp build",
     "build:app": "npm run gulp build:app",
@@ -21,26 +21,21 @@
     "dist:mac": "npm run gulp package:mac",
     "dist:linux": "npm run gulp package:linux",
     "dist:portable": "npm run gulp package:portable",
-    "publish:travis": "npm run gulp package:travis",
-    "publish:appveyor": "npm run gulp package:appveyor",
+    "publish:travis": "npm run gulp publish:travis",
+    "publish:appveyor": "npm run gulp publish:appveyor",
     "___HELPERS___": "Utils for common things I need",
     "preinstall": "cd build && npm install",
     "reinstall": "npm run uninstall && npm install",
     "uninstall": "rm -rf ./node_modules/**/* && rm -rf ./build/node_modules/**/*",
     "___SECONDARY___": "These only get used by other scripts",
-    "backend": "npm run env:dev electron ./shells/electron",
-    "frontend": "webpack-dev-server --env=development --watch --debug --color --config tools/webpack",
-    "frontend:preview": "webpack-dev-server --env=preview --watch --debug --color --config tools/webpack",
-    "mocha": "cross-env NODE_PATH=./build/node_modules mocha --opts tools/mocha/mocha.opts --",
-    "gulp": "gulp --gulpfile tools/gulp --cwd . --color --",
+    "electron": "npm run env:dev electron ./electron",
+    "app": "webpack-dev-server --env=development --watch --debug --color --config config/webpack",
+    "app:preview": "webpack-dev-server --env=preview --watch --debug --color --config config/webpack",
+    "mocha": "cross-env NODE_PATH=./build/node_modules mocha --opts config/mocha/mocha.opts --",
+    "gulp": "gulp --gulpfile config/gulp --cwd . --color --",
     "env:dev": "cross-env NODE_ENV=development NODE_PATH=./build/node_modules",
     "env:test": "cross-env NODE_PATH=./build/node_modules",
-    "env:prod": "cross-env NODE_ENV=production",
-    "___REFACTOR___": "TODOs:",
-    "OLDprepreview": "npm run --silent gulp --silent ensureAppIsBuilt",
-    "OLDpreview": "electron ./build/",
-    "OLDbundle:ui": "npm run gulp bundle:ui",
-    "OLDbundle:electron": "webpack --env=electron --color --config tools/webpack"
+    "env:prod": "cross-env NODE_ENV=production"
   },
   "build": {
     "files": [
